fixed data model

+++ sql/utils/run-db.sh
@@ -7,6 +7,14 @@ shutdown() {
 	pg_ctl -D "$basedir/db" stop
 }
 
+error() {
+    echo "$1" >&2
+    exit 1
+}
+
+[[ -d "$basedir/db" ]] || error "No directory: '$basedir/db'"
+[[ -d "$basedir/db-logs" ]] || error "No directory: '$basedir/db-logs'"
+
 trap "trap - SIGTERM && shutdown && exit 1 && kill -- -$$" SIGINT SIGTERM exit
 
 pg_ctl -D "$basedir/db" -l "$basedir/db-logs/$(date).log" start
